Fix test ad2 on Win32.

+++ test/ad2.jl
@@ -12,4 +12,4 @@ using Base.Test
 end
 
 # Run HMC with chunk_size=1
-chain = sample(ad_test2, Dict(:x=>[1.5 2.0]), HMC(100, 0.1, 1), Int64(1))
+chain = sample(ad_test2, Dict(:x=>[1.5 2.0]), HMC(100, 0.1, 1), 1)
